Update dependencies and add icons
Updated the 'Cargo.toml' and 'Cargo.lock' files to include the 'strum' and 'strum_macros' dependencies, replacing the commented styled and stylist modules.

In 'main.rs', refactored the Component and Icon components for better code organization and introduced use of the 'strum' package.

In 'lib.rs', added a new LucidIconType enum bundled with a display trait and conversion of enum variants to file names, which enhances the flexibility in using different icons.