Adjust icon hover text position for better visibility
Update the position of the icon hover text to the middle, in order to increase visibility. The previous location at the top had usability issues as the text was sometimes not visible to the users. This CSS change helps to improve user experience, especially for new users who are unfamiliar with the icon designations.