Add crc32fast and flate2 packages to manage data compression
Two packages `crc32fast 1.3.2` and `flate2 1.0.27` have been included in the Cargo.lock file.
These new additions are used to handle data compression, particularly in marshalling and unmarshalling of data.
Additional changes are embedded in the `generate_lucide.rs` and `lucide_icon_data.rs` which reflect the usage of these new packages.
Functions for compressing and decompressing are added, and updates are made on how the SVGs are stored and retrieved.
This change improves the performance of the application by reducing the size of the SVG contents and their corresponding load times.