"Added icon metadata output to the icon generation script"
The script for generating icons has been modified to also output metadata related to each icon, including categories, tags, and contributors. This information will be added as properties to each enum value in the output enum. Additionally, the writing of static consts for each icon was commented out. This could improve usability and traceability of the icons by making information such as original authorship easily available.