Refactor Icon component for reactive properties
Updated the Icon component to use reactive properties with the help of MaybeSignal. All static properties like class, size, etc., are now wrapped in MaybeSignal. This allows us to react to changes in these properties and update the component on the fly. Also made sure to make the Glyph trait Cloneable to ensure smooth operation. Furthermore, Github link to the project has been added in the app for easier access.