Refactor Icon component and switch to string constants
This commit revises the Icon and ThemeToggle components. All default symbolic constants in the Icon component have been switched to static strings for consistency. The "size" and "stroke_width" properties are now also retrieved as strings, affecting the generation of the SVG element. Adjustments were also made in the styling of the GitHub Icon link in the main file.

Furthermore, the constant THEME_KEY in the ThemeToggle component was renamed to DARK_MODE for more appropriate terminology. This resulted in changes in getting and setting the local storage items. Lastly, the size attribute was removed from the Icon instance in ThemeToggle, now defaulting to the constant in the Icon component itself.