Improve icon search in demo by allowing multiple words. Update SVG data for multiple icons
Updated the SVG data represented as a string in the 'LucideGlyph' enum for several icons. This change was due to an update in the source SVG files, to ensure that the icons rendering in the app matches the source files exactly. Updated icons include 'Accessibility', 'ActivitySquare', and 'Album', among others. The change does not affect the functionality of the app, only the appearance of the icons.