Update the path and imports for lucide icons
Renamed the lucide_icons directory and its corresponding path in Cargo.toml to lepticons. Also, replaced all the imports of lucide_icons to lepticons throughout the codebase to reflect the changes made in the directory structure. This change was required to improve the module naming convention and make it more identifiable and standard.