Update GitHub link and clean up workflow file
Updated the hyperlink pointing to the GitHub repo for lucid icons and removed an unnecessary line in the preview workflow file. The hyperlink was modified to reflect the accurate repository name i.e., 'lepticons'. The redundant line in the workflow file was creating visual clutter and therefore removed for a cleaner look.