Upgrade dependencies and add new icon count logging
Package dependencies across the codebase have been upgraded to their latest versions for optimisation purposes and to incorporate any newly added features or bug fixes present in the updated packages. A new print line has been added to the code generator that prints the count of newly added icons onto the console. This new print line aids in the quick identification and confirmation of the number of icons added during that particular update.