Add vercel.json to disable deployments
This commit introduces a new vercel.json configuration file. In this configuration file, the "deploymentEnabled" option for both the "master" and "develop" branches is set to false. This was done to prevent automatic deployments from these branches when changes are pushed. The teams can now manually handle deployments when necessary.