Add routing to leptos project
Added a routing functionality to the leptos project. This was achieved by extracting the content of the main application into a separate component that serves as the default route and providing a fallback route for non-existent paths.

Switching to a routing architecture allows the application to grow in complexity by adding more routes/components without disrupting the existing codebase. The redirecting of non-existent paths to a "404 Page not found" component enhances user experience by giving clear feedback when navigating to a non-existent part of the application. The addition of leptos_router to the dependencies aids in managing the different views of the application.