Refactor header to common component
The header presentation code in LicenseView was extracted to a new CommonHeader component for code reusability and better modularity. This ensures that any changes to the header can be easily and consistently implemented throughout the application while adhering to the DRY (Don't Repeat Yourself) principle. This commit does not introduce any changes to the functionality or the visual presentation of the header.